rename the absolute strategies
