+++ editor/src/components/canvas/canvas-strategies/absolute-resize-bounding-box-strategy.spec.tsx
@@ -18,7 +18,7 @@ import {
 import { pickCanvasStateFromEditorState } from './canvas-strategies'
 import { StrategyState } from './interaction-state'
 import { createMouseInteractionForTests } from './interaction-state.test-utils'
-import { multiselectAbsoluteResizeStrategy } from './multiselect-absolute-resize-strategy'
+import { multiselectAbsoluteResizeStrategy } from './absolute-resize-bounding-box-strategy'
 
 function multiselectResizeElements(
   snippet: string,
@@ -70,7 +70,7 @@ function multiselectResizeElements(
   return foldAndApplyCommands(initialEditor, initialEditor, strategyResult, 'permanent').editorState
 }
 
-describe('Absolute Multiselect Resize Strategy', () => {
+describe('Absolute Resize Bounding Box Strategy', () => {
   it('works with element resized from TL corner', async () => {
     const edgePosition: EdgePosition = { x: 0, y: 0 }
     const snippet = `
