+++ editor/src/components/canvas/canvas-strategies/canvas-strategy-types.ts
@@ -27,8 +27,8 @@ export interface InteractionCanvasState {
 export type CanvasStrategyId =
   | 'ABSOLUTE_MOVE'
   | 'ABSOLUTE_REPARENT'
-  | 'ABSOLUTE_RESIZE'
-  | 'MULTISELECT_ABSOLUTE_RESIZE'
+  | 'ABSOLUTE_RESIZE_DELTA'
+  | 'ABSOLUTE_RESIZE_BOUNDING_BOX'
   | 'KEYBOARD_ABSOLUTE_MOVE'
 
 export interface CanvasStrategy {
